+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Bug #558: Fix `SoftDelete` with initiated custom date (@Tigrov)
 - Enh #564: Clarify `$relations` parameter type in `JoinWith::__construct()` from `array<string|Closure>` to
   `array<string|callable(ActiveQueryInterface):void>` (@vjik)
+- Bug #567: Fix properties with hooks (@Tigrov)
 - Bug #562: Fix `ActiveRecordInterface::upsert()` to prioritize passed associative values during updates (@Tigrov)
 - Bug #561: Fix `ActiveRecordInterface::upsert()` with `$updateProperties = false` (@Tigrov)
 - Bug #550: Relation query should be created by related class, not primary model class (@batyrmastyr)